move farm and syrup types to separate file and add reusable functions for duplicated code
