+++ src/components/RewardSlider/RewardSliderItem.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Box, Typography, Button } from '@material-ui/core';
 import { useHistory } from 'react-router-dom';
 import { makeStyles } from '@material-ui/core/styles';
-import { StakingInfo } from 'state/stake/hooks';
+import { StakingInfo } from 'types';
 import { DoubleCurrencyLogo } from 'components';
 import { ReactComponent as HelpIcon } from 'assets/images/HelpIcon.svg';
 import {
