+++ src/components/StakeSyrupModal/StakeSyrupModal.tsx
@@ -4,7 +4,8 @@ import { makeStyles, useTheme } from '@material-ui/core/styles';
 import { TokenAmount } from '@uniswap/sdk';
 import { TransactionResponse } from '@ethersproject/providers';
 import { CustomModal, ColoredSlider, NumericalInput } from 'components';
-import { useDerivedSyrupInfo, SyrupInfo } from 'state/stake/hooks';
+import { useDerivedSyrupInfo } from 'state/stake/hooks';
+import { SyrupInfo } from 'types';
 import { ReactComponent as CloseIcon } from 'assets/images/CloseIcon.svg';
 import { useTokenBalance } from 'state/wallet/hooks';
 import { useActiveWeb3React } from 'hooks';
